Added colorcolumn to vimrc and changed colorscheme
Added colorcolumn so that we will have a visual marker for the 81st
marker. Now we can keep our code from being too long. Also changed
colorscheme's position so that it was after syntax was enabled and to
customized version of desert256 that changes the redline on col 81 to a
darkgrey one.